requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
click==8.3.0
distro==1.9.0
edge-tts==7.2.3
fastmcp==0.1.0
ffmpeg==1.4
ffmpeg-python==0.2.0
frozenlist==1.8.0
future==1.0.0
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mcp==1.19.0
mdurl==0.1.2
MouseInfo==0.1.3
multidict==6.7.0
pip==24.0
propcache==0.4.1
PyAutoGUI==0.9.54
pydantic==2.12.3
pydantic_core==2.41.4
pydantic-settings==2.11.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.1.1
python-multipart==0.0.20
python3-xlib==0.15
pytweening==1.2.0
referencing==0.37.0
rich==14.2.0
rpds-py==0.28.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
tabulate==0.9.0
typer==0.20.0
typing_extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
